+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-03-30T10:22:25.091Z",
+    "2026-03-31T06:51:13.157Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {